Version 1
Here it is and sorry for the delay. This took me longer than originally
imagined but I wanted to migrate everything to Python and not rely on
any DB/SQL.

This version gets the training data from Yahoo (O,H,L,C for all S&P 500
members from 2008 to 2011) and normalizes them according to the article
[Using Random Patterns to Predict the S&P
500](http://laurelhurstmanagement.wordpress.com/2013/05/09/using-random-patterns-to-predict-the-sp-500/).
It then requests Yahoo again for the trading data (2012). It will then
tally each trading pattern against each training pattern and calculate
success rate.

In a few weeks, I'll update this version with a portfolio tester as done
in the second part of the article. Until then, this shows the crux of
the theory discussed in the article and good starting point for any
subsequent efforts.

Of course, I make no guarantee that the code here and in the article are
accurate and certainly wouldn't
recommend trading off of it. Also the numbers I get using the new Python
code and Yahoo data doesn't exactly match the SQL/kinetick and original
code. Let me know if you see any bugs, holes, omissions or have good
ideas.

Thanks!
Manuel Amunategui